Update react-scripts to 2.1.5 (the latest)
This fixes the "warning  Definition for rule 'jsx-a11y/href-no-hash'
was not found  jsx-a11y/href-no-hash" logspam when node starts.

(The issue is that eslint-plugin-jsx-a11y is not compatible
with the version of eslint installed by the old react-scripts we
were using.)

This also adds a default "browserslist" entry to package.json,
which is required by the new version of `npm start` included in
the react-scripts update. (npm start will fail without this entry.)

https://www.npmjs.com/package/browserslist