Add new config option so we can unit test st.* functions.
Because the st.* functions are in streamlit/__init__.py, it makes it
extremely difficult to test since with_dg starts up a connection which
we dont want while testing.

Add option, unit test with manually creates the ReportQueue that we can
inspect manually in the unit test.

A pre step for #481