StatusWidget: fade-in/out animations (#615)
Adds a fade-in and fade-out to StatusWidget appearance and disappearance.

StatusWidget now has a top-level `CSSTransition` element, via the [ReactTransitionGroup](https://github.com/reactjs/react-transition-group) library. It wraps the StatusWidget's actual display elements, and adds/removes the CSS classes that perform the animations automatically.

(We should consider using ReactTransitionGroup to control other bits of animation in the code; it's elegant and simple.)

I also snuck in an update to the latest version of typescript and the various @types packages we're using.