"Script changed" modal dialog (#786)
Removes the "Script changed: Rerun?" prompt from the StatusWidget and implements it as a modal dialog instead.

As part of this PR, I pulled in the react-hotkeys library; it seems to be the de facto solution for React apps, judging by its Github activity, and also looks like it'll help solve our other hotkey woes. (Our current hotkey lib, react-keyboard-shortcuts, doesn't have typescript type definitions and so can't be used in `StreamlitDialog.tsx`). 

This PR _doesn't_ change the way hotkeys are handled anywhere else in the app, but that's an obvious next step!

Fixes https://github.com/streamlit/streamlit/issues/678