Rerun request coalescing bug, take 2 (#1025)
This is an embarrassing followup to https://github.com/streamlit/streamlit/pull/1018.

That PR fixed one of the two edge cases involved in Rerun request coalescing. This PR fixes the other one: coalescing a `RerunRequest` with a null `widget_state` into an existing `RerunRequest` with a non-null widget_state. Test is included.

I also snuck in a tiny change to Server and ScriptRunner, to extend State classes from `Enum` instead of `object`, which is Slightly More Correct™.